requirements.txt
wheel
pandas
numpy
matplotlib
seaborn
nltk
pygooglenews
newspaper3k
beautifulsoup4
requests
pathlib
fpdf
transformers
sentence-transformers
torch
streamlit
spacy
rake-nltk
openpyxl